Split out a file of view models to make code easier to navigate
https://trello.com/c/lN9Z51QU/235-tidy-up-mange-course-ui-fromucasviewmodelcs
