+++ src/ui/ViewModels/CourseVariantStatus.cs
@@ -1,0 +1,11 @@
+﻿using GovUk.Education.ManageCourses.ApiClient;
+
+namespace GovUk.Education.ManageCourses.Ui.ViewModels
+{
+    public enum CourseVariantStatus
+    {
+        Running,
+        NotRunning,
+        New
+    };
+}
